+++ R/geno_freqs_scatter.R
@@ -1,7 +1,5 @@
-
-
-#' return a ggplot of observed and expected genotype freqs
-#'
+#' @title Plot of observed and expected genotype freqs
+#' @description return a ggplot of observed and expected genotype frequencies.
 #' @param gfc a tibble like that created by exp_and_obs_geno_freqs()
 #' @param alpha the transparency (alpha) parameter to apply to the points
 #' in the scatterplot. Default is 0.2.